Begin fork of st -> nast.
Not-as-simple terminal: rearchitect st to create the first terminal with
a strong abstraction between terminal rendering and terminal emulating.

The result will include a standalone terminal rendering library
(libnast) that can be used to simulate the graphical output of a
terminal emulator.  libnast will use Cairo and Pango to render text onto
arbitrary Cairo surfaces.

The nast application will use GTK so that running under X11 or under
Wayland is identical.