Undo TLine rendering.
TLine rendering was meant to track logical lines all the time and let
physical lines be an implementation detail.  That made resizing the
terminal trivially easy.  However, most ANSI escape codes only recognize
physical lines, so supporting ANSI escape codes meant inferring which
logical line every random character represented, and that proved neigh
impossible.

Instead, restore the original physical-location based logic, making
RLine the core line structure.  Resizing the terminal will be trickier
and will depend on tracking which lines represent untainted logical
lines that can be redrawn.  Certain character inserting patterns or
things like literal line breaks will prevent two adjacent RLines from
being treated as a single logical line.

This also introduces scrolling to the rendering, so that writing past
the bottom of the screen behaves correctly.  The scroll is purely a
rendering detail; all cursor operations behave as if no scroll history
existed.  For instance, moving the cursor to the top of the window
places the cursor term.rows rows above the bottom of the line buffer,
which is only the top line being rendered when the scroll == 0.

Handling inputs to change the scroll to anything but 0 is not supported.

Finally, DECOM (origin mode) was removed because it seems weird as shit.