add --logger to raw_inputs
This is in preparation for a mass-duplication of xterm outputs, with the
ultimate goal of using masquerading as xterm-256color.

The way that xterm encodes certain keypresses seems to be listed
absolutely nowhere in the terminfo, such as ctrl-home vs
ctrl-shift-home.  So handling those keypresses well seems like we should
pretend to be a known entity and emit everything the known entity emits.

This will probably mean we have to support a bunch of new features on
the receiving side, but I imagine that an imperfect masquerade is still
99% useful.

Additionally, masquerading as xterm-256color will allow us to work
seamlessly over ssh, since TERM=xterm-256color seems to be pre-installed
everywhere.