rewrite keymap.h from scratch
This marks the beginning of an effort to just masquerade as xterm.

Everything "just works" with xterm, and the terminfo to xterm comes
pre-installed in nearly every environment.  We can just do what xterm
does in enough cases that we are practically indistinguishable to a
terminal application, but much more pleasant as a terminal emulator.