test-io.c, io.c, and some ModifyOtherKeys support
The original goal was to add ModifyOtherKeys support, but things got so
bad it became necessary to design test-io.c, which can empirically test
the io reponse of xterm under various conditions, and it can also ensure
that nast behaves identically.

This caused an upset in the keypad-related io handling.  The old io code
tried to special-case keypad code relative to the input method, but that
led to fragile, fragmented, and incorrect io code.

The new plan will be to remove the input method entirely, because it's
too difficult to support.

Additionally, the keymap logic in keymap.h needs to generalize a bit.
I'm going to model appkeypad and appcursor mode as modifiers.

This code is in a halfway state, and contains the best idea so far for
how to support the input method (remember what we would do without the
input method, then try to detect when the input method is passing
keystrokes through unmodified, and use the precalculated output instead).