finish modifyOtherKeys=1 support
Or rather, support enough of the possible options and give up.

Additional tests can be thought up and passed in the future when
somebody cares enough to report an inconsistency.

Based on the complexity of what I'm seeing, I just can't imagine there
are any applications that actually rely on the finer points of this
holy-shit-complicated xterm IO behavior.